Модуль 7 задание 1 + фиксы
