--- js/isCorrect.js
@@ -1,3 +1,0 @@
-export function isCorrect(str, maxLen) {
-  return str.length <= maxLen;
-}
