+++ js/main.js
@@ -1,8 +1,3 @@
-import { getRandom } from './getRandom.js';
-import { generating } from './generating.js';
-import { isCorrect } from './isCorrect.js';
-
-getRandom(500, 100);
-isCorrect('Kotlin', 7);
-generating(24);
-console.log('hjhj');
+import './utils.js';
+import './data.js';
+import './render.js';
