+++ js/utils.js
@@ -1,3 +1,7 @@
 export function getRandom(a, b) {
   return Math.floor(Math.random() * (b - a + 1)) + a;
 }
+export function isCorrect(str, maxLen) {
+  return str.length <= maxLen;
+}
+
